Fix application closing when choosing mode on Windows
When a new sub form is opened, the main form's ``Close()`` method is called, which halts the entire application on Windows. Instead, we create an instance of the sub form, and bind its ``Closed`` event to an event handler that will halt the entire application.

We can then hide the main form, rather than closing it completely, which resolves the issue and allows the app to run as intended.